refactor: 외부 API 요청 코드 분리
